+++ CHANGES.rst
@@ -4,6 +4,8 @@
 - Fixed a bug in `LightCurve.bin()` which caused the method to fail if the
   ``quality`` array has a floating point data type. [#705]
 
+- Resolved an issue which caused ``interact()`` and ``interact_bls()`` to be
+  incompatible with Bokeh v2.0.0. [#716]
 
 
 1.9.0 (2020-02-25)